requirements.txt
chardet==5.2.0
lxml==5.2.1
python-docx==1.1.0
typing_extensions==4.11.0
